# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
      "AccountProfileProjection": {
        "$id": "AccountProfileProjection.v1",
        "additionalProperties": false,
        "properties": {
          "authority": {
            "items": {
              "type": "string"
            },
            "maxItems": 0,
            "type": "array"
          },
          "displayName": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "verifiedEmail": {
            "maxLength": 320,
            "minLength": 3,
            "type": "string"
          },
          "version": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "displayName",
          "verifiedEmail",
          "authority",
          "version"
        ],
        "type": "object"
      },
...
      "ChangeDisplayNameRequest": {
        "$id": "ChangeDisplayNameRequest.v1",
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "expectedVersion": {
            "minimum": 1,
            "type": "integer"
          },
          "idempotencyKey": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "expectedVersion",
          "idempotencyKey"
        ],
        "type": "object"
      },
...
      "CreateNoticeIntentRequest": {
        "$id": "CreateNoticeIntentRequest.v1",
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 4000,
            "minLength": 1,
            "type": "string"
          },
          "causationId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "correlationId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "expectedAuthorityVersion": {
            "minimum": 1,
            "type": "integer"
          },
          "expectedVersion": {
            "const": 0,
            "type": "number"
          },
          "idempotencyKey": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "noticeIntentId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "recipientAccountId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "subject": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "noticeIntentId",
          "recipientAccountId",
          "subject",
          "body",
          "expectedAuthorityVersion",
          "expectedVersion",
          "idempotencyKey",
          "correlationId",
          "causationId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Docket API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/account/display-name": {
      "post": {
        "operationId": "changeDisplayName",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeDisplayNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocketSessionResult"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The current authority is stale."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The response did not match its contract."
          }
        },
        "summary": "Change the authenticated Account holder's governed Docket Display Name.",
        "x-audience": "account-self",
        "x-requirements": [
          "R1-AUTH-001",
          "R1-LIFE-001",
          "R1-PRIV-001"
        ]
      }
    },
    "/v1/account/profile": {
      "get": {
        "operationId": "getAccountProfile",
        "parameters": [
          {
            "in": "query",
            "name": "audience",
            "required": true,
            "schema": {
              "const": "self",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountProfileProjection"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The response did not match its contract."
          }
        },
        "summary": "Return the authenticated Account holder's current Docket profile.",
        "x-audience": "account-self",
        "x-requirements": [
          "R1-AUTH-001",
          "R1-LIFE-001",
          "R1-PRIV-001"
        ]
      }
    },
    "/v1/account/security-history": {
      "get": {
        "operationId": "listAccountSecurityHistory",
        "parameters": [
          {
            "in": "query",
            "name": "audience",
            "required": true,
            "schema": {
              "const": "self",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSecurityHistoryList"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The response did not match its contract."
          }
        },
        "summary": "List the authenticated Account holder's retained Security History.",
        "x-audience": "account-self",
        "x-requirements": [
          "R1-AUTH-001",
          "R1-PRIV-001"
        ]
      }
    },
    "/v1/communications/inbox": {
      "get": {
        "operationId": "readAccessInbox",
        "parameters": [
          {
            "in": "query",
            "name": "recipientAccountId",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadAccessInboxResult"
                }
              }
            },
            "description": "Successful response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The requested recipient is unavailable."
          }
        },
        "summary": "Read only the authenticated recipient's notice inbox",
        "x-audience": "authorized recipient",
        "x-requirements": [
          "R1-PRIV-001"
        ]
      }
    },
    "/v1/communications/notices": {
      "post": {
        "operationId": "createNoticeIntent",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoticeIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNoticeIntentResult"
                }
              }
            },
            "description": "Successful response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The requested recipient is unavailable."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The idempotency key was already used for different input."
          }
        },
        "summary": "Commit an authorized notice intent and outbox event atomically",
        "x-audience": "initiating actor",
        "x-requirements": [
          "R1-CONS-001",
          "R1-MSG-001",
          "R1-PRIV-001"
        ]
      }
    },
...
    "/v1/docket-sessions/revoke": {
      "post": {
        "operationId": "revokeDocketSession",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeDocketSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocketSessionResult"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "Authentication is required."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StableErrorEnvelope"
                }
              }
            },
            "description": "The response did not match its contract."
          }
        },
        "summary": "Revoke one of the authenticated Account holder's Docket Sessions.",
        "x-audience": "account-self",
        "x-requirements": [
          "R1-AUTH-001",
          "R1-PRIV-001"
        ]
      }
    },
...
  }
}